Extract UserNameProvider to plugin-toolkit
